Cleaned up a bit more. Removed GActor class entirely.
